+++ lib/nostr_mail.dart
@@ -10,6 +10,7 @@ export 'src/constants.dart';
 export 'src/models/email.dart' show Email;
 export 'src/models/encrypted_blob.dart';
 export 'src/models/mail_event.dart';
+export 'src/models/private_settings.dart' show PrivateSettings;
 export 'src/services/email_parser.dart' show EmailParser;
 export 'src/utils/event_email_parser.dart' show parseEmailEvent;
 export 'src/exceptions.dart';